backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forgeroom-backend"
version = "0.1.0"
description = "ForgeRoom backend and AI services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.115.0",
  "gitpython>=3.1.43",
  "httpx>=0.27.2",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.5.2",
  "python-dotenv>=1.0.1",
  "sqlalchemy>=2.0.35",
  "uvicorn>=0.30.6",
]

[project.optional-dependencies]
ai = [
  "langchain-google-genai>=2.0.0",
  "langgraph>=0.2.35",
]
test = [
  "pytest>=8.3.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
